+++ core/constants.py
@@ -1,12 +1,11 @@
 SERVICE_NAME = 'great-cms'
 DASHBOARD_URL = '/dashboard/'
 LOGIN_URL = '/login/'
+EXPORT_PLAN_DASHBOARD_URL = '/export-plan/dashboard/'
 
 VIDEO_DURATION_DATA_ATTR_NAME = 'data-v-duration'
 
 BACKLINK_QUERYSTRING_NAME = 'return-link'
-
-SSO_COOKIE_DOMAIN_NAME_KEY = 'sso_cookie_domain_name'
 
 # Define a strict subset of rich-text features only includes linebreaks
 RICHTEXT_FEATURES__MINIMAL = ()