+++ core/wagtail_hooks.py
@@ -10,6 +10,11 @@ from great_components.helpers import add_next
 from wagtail.core import hooks
 from wagtail.core.models import Page
 
+from wagtail.contrib.modeladmin.options import (
+    ModelAdmin,
+    modeladmin_register
+)
+
 from django.urls import reverse
 from django.core.serializers.json import DjangoJSONEncoder
 from django.shortcuts import redirect
@@ -19,6 +24,29 @@ from core import constants, mixins, views, models
 
 SESSION_KEY_LESSON_PAGE_SHOW_GENERIC_CONTENT = 'LESSON_PAGE_SHOW_GENERIC_CONTENT'
 exportplan_templates = ['exportplan/automated_list_page.html', 'exportplan/dashboard_page.html']
+
+
+class CaseStudyAdmin(ModelAdmin):
+    model = models.CaseStudy
+    add_to_settings_menu = False
+    exclude_from_explorer = False
+    menu_icon = 'fa-book'
+    list_display = (
+        '__str__',
+        'associated_hs_code_tags',
+        'associated_country_code_tags'
+    )
+    list_filter = ('hs_code_tags', 'country_code_tags',)
+    search_fields = ('title', 'company_name',)
+
+    def associated_hs_code_tags(self, obj):
+        return [str(x) for x in obj.hs_code_tags.all()]
+
+    def associated_country_code_tags(self, obj):
+        return [str(x) for x in obj.country_code_tags.all()]
+
+
+modeladmin_register(CaseStudyAdmin)
 
 
 @hooks.register('before_serve_page')