+++ react-components/src/components/CaseStudy/CaseStudy.jsx
@@ -18,10 +18,19 @@ const CaseStudy = ({ content: { heading, company, body } }) => {
     }
   }, [isOpen])
 
+  const toggleCaseStudy = () => {
+    setIsOpen(!isOpen)
+    let dataLayer = window.dataLayer || []
+    dataLayer.push({
+      'event':'openCaseStudy',
+      'caseStudy': heading,
+      });
+  }
+
   return (
-    <div className="case-study p-t-m p-b-s m-b-m">
+    <div className="case-study p-t-m p-b-s">
       {isOpen && (
-        <button className="case-study__close" onClick={() => setIsOpen(false)} autoFocus>
+        <button className="case-study__close" onClick={ toggleCaseStudy } autoFocus>
           <i className="fas fa-times"></i>
           <span className="visually-hidden">Close</span>
         </button>
@@ -29,7 +38,7 @@ const CaseStudy = ({ content: { heading, company, body } }) => {
       <div className="case-study__content media-block">
         <i className="fas fa-newspaper" />
         <div>
-      <h3 className="h-m m-b-xs p-0">{heading}</h3>
+          <h3 className="h-m m-b-xs p-0">{heading}</h3>
           {isOpen && (
             <>
               <p className="case-study__company h-s p-0">{company}</p>
@@ -39,7 +48,7 @@ const CaseStudy = ({ content: { heading, company, body } }) => {
         </div>
       </div>
       {!isOpen && (
-        <button className="button button--small button--tertiary case-study__open" onClick={() => setIsOpen(true)}>
+        <button className="button button--small button--tertiary case-study__open" onClick={toggleCaseStudy}>
           Open case study
         </button>
       )}